build(arduino): Check for compatible hardware at runtime.
Before CRSF for Arduino can be initialised, the library checks whether-or-not it has been flashed onto compatible hardware.
If the hardware is compatible, CRSF for Arduino will initialise & carry on with normal operation.
If CRSF for Arduino is flashed onto incompatible hardware, it will halt the setup process indefinitely.
If the ```CRSF_DEBUG``` flag is enabled at compile time, the name of the development board is visible in the Serial Monitor. If the development board is incompatible, you will see "Unknown device" in the Serial Monitor.