feat(devboard): Add initial support for Teensy 3.2
This adds preliminary support for the Teensy 3.2 development board.

PlatformIO sees Teensy 3.1 and Teensy 3.2 as the same board, whereas the Arduino IDE sees them as two separate boards. This commit prevents the Compatibility Table from causing a false negative when CRSF for Arduino is compiled for Teensy 3.2 in the Arduino IDE.