refactor(compatibility table): Permissive Incompatibility
CRSF for Arduino will now attempt to run, if it's flashed on an unknown development board and/or chipset.  
Additionally, you now have instructions in the Serial Monitor that provide you with information on how to go about requesting support for your development board.

This brings in two configuration defines in `CFA_Config.hpp`:  

| Pre-processor define | Description |
| --- | --- |
| `CRSF_DEBUG_ENABLED` | When set to `1`, this allows debug outputs from CRSF for Arduino to appear in the Serial Monitor. |
| `CRSF_DEBUG_ENABLE_COMPATIBILITY_TABLE_OUTPUT` |  If the Compatibility Table pings you for something, set _this_ to `1` to find out why. |