chore(library): Deprecate existing files-and-folders structure
The current file-and-folder structure has proven itself to be problematic when CRSF for Arduino is used as a PlatformIO dependency. This marks the beginning of a re-write of the entire file-and-folder structure to improve compatibility with PlatformIO whilst maintaining compatibility with the Arduino IDE.

While efforts are underway to minimise the possibility of CRSF for Arduino's functionality being broken, there are no guarantees that this re-factor will not introduce any broken code.