fix(arduino): Arduino IDE build failure (#94)
Some typos in `library.properties` completely broke CRSF for Arduino's compatibility with the Arduino IDE. This has been fixed.  
Along with some tidy-up of the badges in `README.md` and the Build Status badges are now working the way thar supposed to, as well.

**NB:** The Arduino build status indicates `failing`, because I am still experiencing issues with the Arduino CI itself. This _does not_ reflect how CRSF for Arduino will build in the Arduino IDE itself.  
CRSF for Arduino was locally tested on the Arduino IDE v2.3.2 on an Arduino Nano ESP32 target, and the build passed with impunity.