chore(library): :fire: Drop Semantic Version and Date from every individual header and source file
Currently, with over 28 files, CRSF for Arduino is incredibly tedious to re-base because of these two tags in the boiler plate. In addition to this, the version numbering and dates are already tracked in `CFA_Config.hpp`. Therefore, it is pointless to keep tracking CRSF for Arduino's overall version and date in every individual file. At this point, it's work for work's sake, and it's unnecessarily tedious to do it... even more so as CRSF for Arduino grows and more headers and source files are added over time. The juice is not worth the squeeze to track CRSF for Arduino's version and date in every individual file, when it's more efficient to do it in the config and properties files.