chore(platformio): Move all PlatformIO configuration files to src/build/scripts and src/build/targets
This moves all of PlatformIO's files (including the main platformio.ini and all target-specific .ini files) to a new directory structure under src/build/scripts and src/build/targets.
This change is made to simplify the build and check process, making it easier to manage and maintain the configuration files.

**Breaking Changes:**

- The paths to the PlatformIO configuration files have changed.
  - The main `platformio.ini` file is now located at `src/build/scripts/platformio.ini`.
  - Target-specific configuration files are now located in `src/build/targets/`.
  - The previous paths to these files will no longer work.