ci(quality control): Simplify QC, build, and check process (#156)
## Changes

This is a Python rewrite of the old Bash scripts used to start the Quality Control CI pipeline.

## Breaking changes

This pull request breaks the existing QC and vulnerability scanning CI workflows, because I changed the directories around to be more aligned with CFA's development container and I moved the main `platformio.ini` file away into a folder that houses the build script to run everything from there. So, the existing CI is expecting that `platformio.ini` file to still be present in CFA's root.