fix(serial receiver interface): :rotating_light: Explicitly add default constructor and destructor to `crsf_broadcast_frame_u`
This was being implicitly deleted by the compiler, causing the union to be ill-formed. I am not out of the woods with this yet, as I need to make it type-safe too. But at least this headache is done-and-dusted for now. Honestly, in all my years of writing C++ firmware since 2013, I have _never_ encountered a bug like this until now. It actually had me going "What the fuck?" and it took me a few days to figure it out, but I got there in the end. =^/.^=